pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lore-sdk"
version = "0.1.1"
description = "Cross-agent memory library"
readme = "README.md"
license = "MIT"
authors = [{ name = "Amit Paz", email = "amit.paz@gmail.com" }]
keywords = ["agent", "memory", "lessons", "lore"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.9"
dependencies = [
    "python-ulid>=2.0.0",
    "numpy>=1.21.0",
    "onnxruntime>=1.14.0",
    "tokenizers>=0.13.0",
]

[project.optional-dependencies]
server = [
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "asyncpg>=0.28.0",
    "pydantic>=2.0.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
    "ruff>=0.1.0",
]

[project.scripts]
lore = "lore.cli:main"

[project.urls]
Homepage = "https://github.com/amitpaz1/lore"
Repository = "https://github.com/amitpaz1/lore"
Issues = "https://github.com/amitpaz1/lore/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/lore"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
